+++ src/views/pages/plugins/List/List.vue
@@ -1,22 +1,18 @@
 <template>
     <div>
-        <AssetListPageWrap class="plugin-list-container">
-            <template v-slot:table-section>
-                <TableDisplay
-                    asset-type="plugin"
-                    class="table-section"
-                    :pagination-type="'scroll'"
-                    :page-size="20"
-                    @on-clicked-row="onRowClick"
-                >
-                    <template v-slot:table-title>
-                        <h1 class="section-title">
-                            {{ $t('title_plugin_manager') }}
-                        </h1>
-                    </template>
-                </TableDisplay>
+        <TableDisplay
+            asset-type="plugin"
+            class="table-section"
+            :pagination-type="'scroll'"
+            :page-size="20"
+            @on-clicked-row="onRowClick"
+        >
+            <template v-slot:table-title>
+                <h1 class="section-title">
+                    {{ $t('title_plugin_manager') }}
+                </h1>
             </template>
-        </AssetListPageWrap>
+        </TableDisplay>
     </div>
 </template>