+++ src/views/pages/plugins/PageWrapper/PageWrapper.vue
@@ -1,7 +1,10 @@
 <template>
     <div v-if="!!selectedPlugin" class="plugin-page-container">
         <div v-if="pluginComponent !== null">
-            <Component :is="pluginComponent" @prepared="onTransactionPrepared"></Component>
+            <Component
+                :is="pluginComponent"
+                @prepared="onTransactionPrepared"
+                @account-request="onAccountRequestIntercepted"></Component>
         </div>
         <div v-else>
             <span>This plugin does not include a page.</span>
@@ -15,6 +18,22 @@
             @success="onConfirmationSuccess"
             @error="onConfirmationError"
             @close="onConfirmationCancel"
+        />
+
+        <ModalConfirm
+            v-model="showConfirmAccountRequestModal"
+            :title="$t('modal_title_confirm_account_request')"
+            :message="$t('modal_text_confirm_account_request')"
+            :use-bigger-font="true"
+            @confirmed="onAccountRequestAccepted"
+        />
+
+        <ModalFormSubAccountCreation
+            v-if="showSubAccountFormModal"
+            :visible="showSubAccountFormModal"
+            :should-list="false"
+            @close="showSubAccountFormModal = false"
+            @account-added="onAccountRequestCompleted"
         />
     </div>
 </template>