+++ CHANGELOG.md
@@ -3,6 +3,17 @@ All notable changes to this project will be documented in this file.
 
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.0][1.1.0] - 21-Sep-2021
+
+#### Added
+
+- Added plugin manager for Plug & Play dApps
+- Added script `npm run plugins:install`
+- Added `AppPluginManager` in Electron bundler
+- Added plugins dual communication protocol
+- Added `$pluginBus` with plugin-only events
+- Added AccountModel.isListedAccount
+
 ## [1.0.0][1.0.0] - 03-May-2021
 
 #### Changed